+++ tests/conftest.py
@@ -1,3 +1,5 @@
+import random
+import string
 from abc import ABC, ABCMeta
 from functools import wraps
 from pathlib import Path
@@ -25,6 +27,13 @@ EXPECTED = HERE / "plots/groundtruth"
 ACTUAL = HERE / "plots/generated"
 TOL = 70
 DPI = 40
+
+RNG = np.random.default_rng(seed=0)
+DATA_LEN = 100
+
+
+def pytest_configure(config):
+    config.DATA_LEN = 100
 
 
 @pytest.fixture
@@ -114,6 +123,43 @@ def image():
 def labels():
     blobs, _ = _get_blobs_galaxy()
     return blobs
+
+
+@pytest.fixture
+def prepare_cont_test_data():
+    x_vec = RNG.random(DATA_LEN)
+    y_vec = RNG.random(DATA_LEN)
+    color_vec = RNG.random(DATA_LEN)
+
+    x_data = {"vec": x_vec}
+    y_data = {"vec": y_vec}
+    color_data = {"vec": color_vec}
+
+    x_label = generate_random_string(10)
+    y_label = generate_random_string(10)
+    color_label = generate_random_string(10)
+    return x_data, y_data, color_data, x_label, y_label, color_label
+
+
+@pytest.fixture
+def prepare_disc_test_data():
+    x_vec = RNG.random(DATA_LEN)
+    y_vec = RNG.random(DATA_LEN)
+    color_vec = np.zeros(DATA_LEN).astype(int)
+
+    x_data = {"vec": x_vec}
+    y_data = {"vec": y_vec}
+    color_data = {"vec": color_vec, "labels": ["a"]}
+
+    x_label = generate_random_string(10)
+    y_label = generate_random_string(10)
+    color_label = generate_random_string(10)
+    return x_data, y_data, color_data, x_label, y_label, color_label
+
+
+def generate_random_string(length):
+    letters = string.ascii_letters  # Includes both lowercase and uppercase letters
+    return "".join(random.choice(letters) for i in range(length))
 
 
 def _get_blobs_galaxy() -> Tuple[NDArrayA, NDArrayA]:
